backend/requirements/base.txt
Django==5.0.4
djangorestframework==3.15.1
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
redis==5.0.4
celery==5.3.6
django-celery-beat==2.6.0
python-decouple==3.8
gunicorn==22.0.0